+++ lib/stats/lighthouse.ts
@@ -1,24 +1,16 @@
-import * as http from 'http'
 import * as path from 'path'
-import { fork } from 'child_process'
-import * as dotenv from 'dotenv'
+import * as Lighthouse from 'lighthouse'
+import { ChromeLauncher } from 'lighthouse/lighthouse-cli/chrome-launcher'
 import CONFIG from '../config/config'
-import { getPort } from '../server/server'
-import { waitForConnection } from '../util/network'
 import { log, logError, prettyJson } from '../util/log'
 import { BuildMetrics } from './'
+import { runStatsOnServer } from './server'
 import { readFile } from '../util/fs'
 import { delay, formatTimeMs } from '../util/time'
 
-import * as Lighthouse from 'lighthouse'
-import { ChromeLauncher } from 'lighthouse/lighthouse-cli/chrome-launcher'
+const { HAS_SERVER } = CONFIG
 
-dotenv.config()
-
-const { SERVER_OUTPUT, HAS_SERVER } = CONFIG
-const port = getPort()
-
-const runLighthouse = async () => {
+export const runLighthouse = async (url: string) => {
     const launcher = new ChromeLauncher({
         port: 9222,
         autoSelectChrome: true,
@@ -46,8 +38,11 @@ const runLighthouse = async () => {
 
     try {
         const results = await Lighthouse(
-            `http://localhost:${port}`,
-            { output: 'json', port: 9222 },
+            url,
+            {
+                output: 'json',
+                port: 9222,
+            },
             config,
         )
 
@@ -68,16 +63,6 @@ const runLighthouse = async () => {
     }
 }
 
-const loadPage = () => (
-    new Promise<void>((resolve, reject) => {
-        const request = http.get('http://localhost:' + port, (res) => {
-            res.on('data', () => {})
-            res.on('end', () => resolve())
-        })
-        request.on('error', (err) => reject(err))
-    })
-)
-
 const lighthouseStats = async (): Promise<BuildMetrics> => {
 
     if (!HAS_SERVER) {
@@ -87,45 +72,27 @@ const lighthouseStats = async (): Promise<BuildMetrics> => {
 
     log('Measuring lighthouse performance metrics...')
 
+    const stats: BuildMetrics = {}
+
     try {
-        const serverEntryFile = path.resolve(SERVER_OUTPUT, 'server.js')
+        await runStatsOnServer(async (page: string, _urlPath: string, url: string) => {
+            const lighthouseResult = await runLighthouse(url)
 
-        const devServer = fork(
-            serverEntryFile,
-            [],
-            {
-                env: process.env,
-                silent: true,
-            },
-        )
+            const addLighthouseValue = (lighthouseKey: string, statsKey: string) => {
+                const result = lighthouseResult
+                    && lighthouseResult.audits
+                    && lighthouseResult.audits[lighthouseKey]
+                    && (lighthouseResult.audits[lighthouseKey].rawValue as number)
 
-        await waitForConnection(port)
+                if (result !== undefined) {
+                    stats[`${page}_${statsKey}`] = formatTimeMs(result)
+                }
+            }
 
-        // warm-up
-        await Promise.all([
-            loadPage(),
-            loadPage(),
-            loadPage(),
-        ])
-
-        const lighthouseResult = await runLighthouse()
-
-        const getLighthouseValue = (key: string) => (
-            formatTimeMs((
-                lighthouseResult
-                && lighthouseResult.audits
-                && lighthouseResult.audits[key]
-                && lighthouseResult.audits[key].rawValue as number
-            ) || 0)
-        )
-
-        devServer.kill()
-
-        const stats: BuildMetrics = {
-            first_meaningful_paint: getLighthouseValue('first-meaningful-paint'),
-            speed_index: getLighthouseValue('speed-index-metric'),
-            time_to_interactive: getLighthouseValue('time-to-interactive'),
-        }
+            addLighthouseValue('first-meaningful-paint', 'first_meaningful_paint')
+            addLighthouseValue('speed-index-metric', 'speed_index')
+            addLighthouseValue('time-to-interactive', 'time_to_interactive')
+        })
 
         log(`Lighthouse stats: ${prettyJson(stats)}`)