+++ lib/stats/ssr-stats.ts
@@ -1,44 +1,12 @@
-import * as path from 'path'
-import * as http from 'http'
-import { fork } from 'child_process'
-import * as dotenv from 'dotenv'
 import CONFIG from '../config/config'
-import { getPort } from '../server/server'
-import { waitForConnection } from '../util/network'
 import { log, logError, prettyJson } from '../util/log'
 import { BuildMetrics } from './'
 import { formatFileSize } from '../util/fs'
-import { getTimeMs, formatTimeMs } from '../util/time'
+import { formatTimeMs } from '../util/time'
 import { getSequenceAverage } from '../util/math'
+import { runStatsOnServer, loadSSRPage } from './server'
 
-dotenv.config()
-
-const { SERVER_OUTPUT, HAS_SERVER } = CONFIG
-const port = getPort()
-
-interface SSRStats {
-    size: number
-    time: number
-}
-
-const loadSSRPage = () => (
-    new Promise<SSRStats>((resolve, reject) => {
-        const startTime = getTimeMs()
-        const request = http.get('http://localhost:' + port, (res) => {
-            res.setEncoding('utf8')
-
-            let size = 0
-            res.on('data', (chunk) => size += chunk.length)
-
-            res.on('end', () => {
-                const time = getTimeMs() - startTime
-                log('SSR load took ' + time + ' ms')
-                resolve({ size, time })
-            })
-        })
-        request.on('error', (err) => reject(err))
-    })
-)
+const { HAS_SERVER } = CONFIG
 
 const ssrStats = async () => {
 
@@ -49,36 +17,18 @@ const ssrStats = async () => {
 
     log('Measuring SSR load times...')
 
+    const stats: BuildMetrics = {}
+
     try {
-        const serverEntryFile = path.resolve(SERVER_OUTPUT, 'server.js')
+        await runStatsOnServer(async (page: string, url: string) => {
+            const { size } = await loadSSRPage(url)
 
-        const devServer = fork(
-            serverEntryFile,
-            [],
-            {
-                env: process.env,
-                silent: true,
-            },
-        )
+            const getTime = () => loadSSRPage(url).then((result) => result.time)
+            const time = await getSequenceAverage(getTime, 5)
 
-        await waitForConnection(port)
-
-        // warm-up
-        const { size } = await loadSSRPage()
-        await loadSSRPage()
-        await loadSSRPage()
-
-        const getTime = () => loadSSRPage()
-            .then((result) => result.time)
-
-        const time = await getSequenceAverage(getTime, 5)
-
-        devServer.kill()
-
-        const stats: BuildMetrics = {
-            ssr_document_size: formatFileSize(size),
-            ssr_loadtime: formatTimeMs(time),
-        }
+            stats[`${page}_ssr_document_size`] = formatFileSize(size)
+            stats[`${page}_ssr_loadtime`] = formatTimeMs(time)
+        })
 
         log(`SSR stats: ${prettyJson(stats)}`)