+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add `debug` environment for `pwt build server debug`
 * Add `pwt start debug` and `START_DEBUG_MODE` environment flag
+* Add `WATCH_IGNORE` configuration
 
 ## v0.1.3 (2017-05-01)